updated files referencing the relocated container files
