+++ client/src/components/ViewProjectRedux.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Container } from 'react-bootstrap'
 import { connect } from 'react-redux';
 import './Style.scss';
-import ViewAProject from './containers/ViewAProject';
+import ViewAProject from './ViewAProjectContainer';
